# Mimir query engine: add support for aggregations with `without`, and resolve potential for hash collisions