… CHANGELOG.md
… pkg/streamingpromql/README.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/operators/aggregation_test.go
… pkg/streamingpromql/testdata/ours/aggregators.test
… pkg/streamingpromql/testdata/upstream/aggregators.test