@charleskorn 2024-07-16
> We should add some test cases where duplicate groups exist. For example:
> 
> `sum without (env, __name__, __abc__, env) metric{}`
> 
> (I used `__name__` explicitly since it is added. I also used `__abc__` since it'll be sorted to before `__name__`).

Good idea, added in fff95c4.